requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2018.1.18
chardet==3.0.4
click==6.7
cycler==0.10.0
Flask==0.12.2
Flask-Script==2.0.6
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
idna==2.6
itsdangerous==0.24
jieba==0.39
Jinja2==2.10
kiwisolver==1.0.1
lxml==4.2.0
MarkupSafe==1.0
matplotlib==2.2.2
numpy==1.14.2
Pillow==5.0.0
pyparsing==2.2.0
python-dateutil==2.7.0
python-docx==0.8.6
pytz==2018.3
regex==2018.2.21
requests==2.18.4
scipy==1.0.0
six==1.11.0
urllib3==1.22
Werkzeug==0.14.1
wordcloud==1.4.1
WTForms==2.1
